// src/models/UserInfo-model.ts
export class UserInfo {
   id: string | undefined;
   Name: string | undefined;
   age: string | undefined;
   email : string | undefined;
   password : string | undefined;
   profilephotolink : string | undefined;
   gender: string | undefined;
   height: string | undefined;
   weight: string | undefined;
   bio: string | undefined;
   nationality: string | undefined;
   instaprofile: string | undefined;
   profession1: string | undefined;
   profession2: string | undefined;
} 

// SR884558516